+++ starboard/android/shared/experimental_features_internal.h
@@ -15,14 +15,14 @@
 #ifndef STARBOARD_ANDROID_SHARED_EXPERIMENTAL_FEATURES_INTERNAL_H_
 #define STARBOARD_ANDROID_SHARED_EXPERIMENTAL_FEATURES_INTERNAL_H_
 
+#include "starboard/extension/experimental_features.h"
 #include "starboard/shared/starboard/player/filter/experimental_features.h"
 
 namespace starboard::android::shared {
 
 // Sets the experimental features for the current thread.
 void SetExperimentalFeaturesForCurrentThread(
-    const starboard::shared::starboard::player::filter::ExperimentalFeatures&
-        experimental_features);
+    const StarboardExtensionExperimentalFeatures* extension_features);
 
 // Gets the experimental features for the current thread.
 const starboard::shared::starboard::player::filter::ExperimentalFeatures&
